HWKALERTS-54 : Alerts spinning the logs with errors
WIP: Refactor alerts service into two services/beans, a pooled bean that
handles typical client interactions and queries, and a singleton devoted to
executing the engine.
- Don't use a singleton for handling potentially frequent and concurrent
  alert fetches and other client requests.
- Isolate engine exection
- Better handle the "pendingData" list wrt concurrency and perf.

Question: Do we really need/want to expose the "reload" endpoints in AlertsHandler?
