+++ hawkular-alerts-engine/src/main/java/org/hawkular/alerts/engine/impl/AlertsEngineImpl.java
@@ -1,0 +1,426 @@
+/*
+ * Copyright 2015 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.hawkular.alerts.engine.impl;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.Timer;
+import java.util.TimerTask;
+
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
+import javax.ejb.EJB;
+import javax.ejb.Singleton;
+
+import org.hawkular.alerts.api.model.condition.Alert;
+import org.hawkular.alerts.api.model.condition.Condition;
+import org.hawkular.alerts.api.model.condition.ConditionEval;
+import org.hawkular.alerts.api.model.dampening.Dampening;
+import org.hawkular.alerts.api.model.data.Data;
+import org.hawkular.alerts.api.model.trigger.Trigger;
+import org.hawkular.alerts.api.services.ActionsService;
+import org.hawkular.alerts.api.services.AlertsService;
+import org.hawkular.alerts.api.services.DefinitionsService;
+import org.hawkular.alerts.engine.log.MsgLogger;
+import org.hawkular.alerts.engine.rules.RulesEngine;
+import org.hawkular.alerts.engine.service.AlertsEngine;
+import org.jboss.logging.Logger;
+
+import com.datastax.driver.core.Session;
+import com.google.gson.GsonBuilder;
+
+/**
+ * Cassandra implementation for {@link org.hawkular.alerts.api.services.AlertsService}.
+ * This implementation processes data asynchronously using a buffer queue.
+ *
+ * @author Jay Shaughnessy
+ * @author Lucas Ponce
+ */
+@Singleton
+public class AlertsEngineImpl implements AlertsEngine {
+    private final MsgLogger msgLog = MsgLogger.LOGGER;
+    private final Logger log = Logger.getLogger(AlertsEngineImpl.class);
+
+    private static final String ENGINE_DELAY = "hawkular-alerts.engine-delay";
+    private static final String ENGINE_PERIOD = "hawkular-alerts.engine-period";
+    private static final String CASSANDRA_KEYSPACE = "hawkular-alerts.cassandra-keyspace";
+
+    private int delay;
+    private int period;
+
+    private Session session;
+    private String keyspace;
+
+    private final List<Data> pendingData;
+    private final List<Alert> alerts;
+    private final Set<Dampening> pendingTimeouts;
+    private final Map<Trigger, List<Set<ConditionEval>>> autoResolvedTriggers;
+    private final Set<Trigger> disabledTriggers;
+
+    private final Timer wakeUpTimer;
+    private TimerTask rulesTask;
+
+    @EJB
+    RulesEngine rules;
+
+    @EJB
+    DefinitionsService definitions;
+
+    @EJB
+    ActionsService actions;
+
+    @EJB
+    AlertsService alertsService;
+
+    public AlertsEngineImpl() {
+        pendingData = new ArrayList<>();
+        alerts = new ArrayList<>();
+        pendingTimeouts = new HashSet<>();
+        autoResolvedTriggers = new HashMap<>();
+        disabledTriggers = new HashSet<>();
+
+        wakeUpTimer = new Timer("CassAlertsServiceImpl-Timer");
+
+        delay = new Integer(AlertProperties.getProperty(ENGINE_DELAY, "1000"));
+        period = new Integer(AlertProperties.getProperty(ENGINE_PERIOD, "2000"));
+    }
+
+    public RulesEngine getRules() {
+        return rules;
+    }
+
+    public void setRules(RulesEngine rules) {
+        this.rules = rules;
+    }
+
+    public DefinitionsService getDefinitions() {
+        return definitions;
+    }
+
+    public void setDefinitions(DefinitionsService definitions) {
+        this.definitions = definitions;
+    }
+
+    public ActionsService getActions() {
+        return actions;
+    }
+
+    public void setActions(ActionsService actions) {
+        this.actions = actions;
+    }
+
+    public AlertsService getAlertsService() {
+        return alertsService;
+    }
+
+    public void setAlertsService(AlertsService alertsService) {
+        this.alertsService = alertsService;
+    }
+
+    @PostConstruct
+    public void initServices() {
+        try {
+            if (this.keyspace == null) {
+                this.keyspace = AlertProperties.getProperty(CASSANDRA_KEYSPACE, "hawkular_alerts");
+            }
+
+            if (session == null) {
+                session = CassCluster.getSession();
+            }
+
+            GsonBuilder gsonBuilder = new GsonBuilder();
+            gsonBuilder.registerTypeHierarchyAdapter(ConditionEval.class, new GsonAdapter<ConditionEval>());
+
+            reload();
+        } catch (Throwable t) {
+            if (log.isDebugEnabled()) {
+                t.printStackTrace();
+            }
+            msgLog.errorCannotInitializeAlertsService(t.getMessage());
+        }
+    }
+
+    @PreDestroy
+    public void shutdown() {
+        if (session != null) {
+            session.close();
+        }
+    }
+
+    public void clear() {
+        rulesTask.cancel();
+
+        rules.clear();
+
+        pendingData.clear();
+        alerts.clear();
+        pendingTimeouts.clear();
+        autoResolvedTriggers.clear();
+        disabledTriggers.clear();
+
+        rulesTask = new RulesInvoker();
+        wakeUpTimer.schedule(rulesTask, delay, period);
+    }
+
+    @Override
+    public void reload() {
+        rules.reset();
+        if (rulesTask != null) {
+            rulesTask.cancel();
+        }
+
+        Collection<Trigger> triggers = null;
+        try {
+            triggers = definitions.getAllTriggers();
+        } catch (Exception e) {
+            log.debugf(e.getMessage(), e);
+            msgLog.errorDefinitionsService("Triggers", e.getMessage());
+        }
+        if (triggers != null && !triggers.isEmpty()) {
+            triggers.stream().filter(Trigger::isEnabled).forEach(this::reloadTrigger);
+        }
+
+        rules.addGlobal("log", log);
+        rules.addGlobal("actions", actions);
+        rules.addGlobal("alerts", alerts);
+        rules.addGlobal("pendingTimeouts", pendingTimeouts);
+        rules.addGlobal("autoResolvedTriggers", autoResolvedTriggers);
+        rules.addGlobal("disabledTriggers", disabledTriggers);
+
+        rulesTask = new RulesInvoker();
+        wakeUpTimer.schedule(rulesTask, delay, period);
+    }
+
+    public void reloadTrigger(final String tenantId, final String triggerId) {
+        if (isEmpty(tenantId)) {
+            throw new IllegalArgumentException("TenantId must be not null");
+        }
+        if (isEmpty(triggerId)) {
+            throw new IllegalArgumentException("TriggerId must be not null");
+        }
+
+        Trigger trigger = null;
+        try {
+            trigger = definitions.getTrigger(tenantId, triggerId);
+        } catch (Exception e) {
+            log.debugf(e.getMessage(), e);
+            msgLog.errorDefinitionsService("Trigger", e.getMessage());
+        }
+        if (null == trigger) {
+            log.debugf("Trigger not found for triggerId [" + triggerId + "], removing from rulebase if it exists");
+            Trigger doomedTrigger = new Trigger(triggerId, "doomed");
+            removeTrigger(doomedTrigger);
+            return;
+        }
+
+        reloadTrigger(trigger);
+    }
+
+    private void reloadTrigger(Trigger trigger) {
+        if (null == trigger) {
+            throw new IllegalArgumentException("Trigger must be not null");
+        }
+
+        // Look for the Trigger in the rules engine, if it is there then remove everything about it
+        removeTrigger(trigger);
+
+        if (trigger.isEnabled()) {
+            try {
+                Collection<Condition> conditionSet = definitions.getTriggerConditions(trigger.getTenantId(),
+                        trigger.getId(), null);
+                Collection<Dampening> dampenings = definitions.getTriggerDampenings(trigger.getTenantId(),
+                        trigger.getId(), null);
+
+                rules.addFact(trigger);
+                rules.addFacts(conditionSet);
+                if (!dampenings.isEmpty()) {
+                    rules.addFacts(dampenings);
+                }
+            } catch (Exception e) {
+                log.debugf(e.getMessage(), e);
+                msgLog.errorDefinitionsService("Conditions/Dampening", e.getMessage());
+            }
+        }
+    }
+
+    private void removeTrigger(Trigger trigger) {
+        if (null != rules.getFact(trigger)) {
+            // First remove the related Trigger facts from the engine
+            rules.removeFact(trigger);
+
+            // then remove everything else.
+            // We may want to do this with rules, because as is, we need to loop through every Fact in
+            // the rules engine doing a relatively slow check.
+            final String triggerId = trigger.getId();
+            rules.removeFacts(t -> {
+                if (t instanceof Dampening) {
+                    return ((Dampening) t).getTriggerId().equals(triggerId);
+                } else if (t instanceof Condition) {
+                    return ((Condition) t).getTriggerId().equals(triggerId);
+                }
+                return false;
+            });
+        } else {
+            log.debugf("Trigger not found. Not removed from rulebase %s", trigger);
+        }
+    }
+
+    @Override
+    public void sendData(Collection<Data> data) {
+        if (data == null) {
+            throw new IllegalArgumentException("Data must be not null");
+        }
+        addPendingData(data);
+    }
+
+    @Override
+    public void sendData(Data data) {
+        if (data == null) {
+            throw new IllegalArgumentException("Data must be not null");
+        }
+        addPendingData(data);
+    }
+
+    private synchronized void addPendingData(Collection<Data> data) {
+        pendingData.addAll(data);
+    }
+
+    private synchronized void addPendingData(Data data) {
+        pendingData.add(data);
+    }
+
+    private synchronized Collection<Data> getAndClearPendingData() {
+        Collection<Data> result = new ArrayList<>(pendingData);
+        pendingData.clear();
+        return result;
+    }
+
+    private class RulesInvoker extends TimerTask {
+        @Override
+        public void run() {
+            int numTimeouts = checkPendingTimeouts();
+
+            if (!pendingData.isEmpty() || numTimeouts > 0) {
+                Collection<Data> newData = getAndClearPendingData();
+
+                log.debugf("Executing rules engine on [%1d] datums and [%2d] dampening timeouts.", newData.size(),
+                        numTimeouts);
+
+                try {
+                    if (newData.isEmpty()) {
+                        rules.fireNoData();
+
+                    } else {
+                        rules.addData(newData);
+                        newData.clear();
+                    }
+
+                    rules.fire();
+                    alertsService.addAlerts(alerts);
+                    alerts.clear();
+                    handleDisabledTriggers();
+                    handleAutoResolvedTriggers();
+
+                } catch (Exception e) {
+                    e.printStackTrace();
+                    log.debugf("Error on rules processing: " + e);
+                    msgLog.errorProcessingRules(e.getMessage());
+                } finally {
+                    alerts.clear();
+                }
+            }
+        }
+
+        private int checkPendingTimeouts() {
+            if (pendingTimeouts.isEmpty()) {
+                return 0;
+            }
+
+            long now = System.currentTimeMillis();
+            Set<Dampening> timeouts = null;
+            for (Dampening d : pendingTimeouts) {
+                if (now < d.getTrueEvalsStartTime() + d.getEvalTimeSetting()) {
+                    continue;
+                }
+
+                d.setSatisfied(true);
+                try {
+                    log.debugf("Dampening Timeout Hit! %s", d.toString());
+                    rules.updateFact(d);
+                    if (null == timeouts) {
+                        timeouts = new HashSet<>();
+                    }
+                    timeouts.add(d);
+                } catch (Exception e) {
+                    log.error("Unable to update Dampening Fact on Timeout! " + d.toString(), e);
+                }
+
+            }
+
+            if (null == timeouts) {
+                return 0;
+            }
+
+            pendingTimeouts.removeAll(timeouts);
+            return timeouts.size();
+        }
+    }
+
+    private void handleDisabledTriggers() {
+        try {
+            for (Trigger t : disabledTriggers) {
+                try {
+                    t.setEnabled(false);
+                    definitions.updateTrigger(t.getTenantId(), t);
+
+                } catch (Exception e) {
+                    log.errorf("Failed to persist updated trigger. Could not autoDisable %s", t);
+                }
+            }
+        } finally {
+            disabledTriggers.clear();
+        }
+    }
+
+    private void handleAutoResolvedTriggers() {
+        try {
+            for (Map.Entry<Trigger, List<Set<ConditionEval>>> entry : autoResolvedTriggers.entrySet()) {
+                Trigger t = entry.getKey();
+                try {
+                    if (t.isAutoResolveAlerts()) {
+                        alertsService.resolveAlertsForTrigger(t.getTenantId(), t.getId(), "AUTO", null,
+                                entry.getValue());
+                    }
+                } catch (Exception e) {
+                    log.errorf("Failed to resolve Alerts. Could not AutoResolve alerts for trigger %s", t);
+                }
+            }
+        } finally {
+            autoResolvedTriggers.clear();
+        }
+    }
+
+    private boolean isEmpty(String s) {
+        return null == s || s.trim().isEmpty();
+    }
+
+}
