+++ hawkular-alerts-engine/src/main/java/org/hawkular/alerts/engine/impl/DbAlertsServiceImpl.java
@@ -27,15 +27,8 @@ import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.EnumSet;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
-import java.util.Timer;
-import java.util.TimerTask;
-import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.CopyOnWriteArraySet;
 
 import javax.annotation.PostConstruct;
 import javax.ejb.EJB;
@@ -45,18 +38,13 @@ import javax.sql.DataSource;
 
 import org.hawkular.alerts.api.model.condition.Alert;
 import org.hawkular.alerts.api.model.condition.Alert.Status;
-import org.hawkular.alerts.api.model.condition.Condition;
 import org.hawkular.alerts.api.model.condition.ConditionEval;
-import org.hawkular.alerts.api.model.dampening.Dampening;
 import org.hawkular.alerts.api.model.data.Data;
 import org.hawkular.alerts.api.model.trigger.Tag;
-import org.hawkular.alerts.api.model.trigger.Trigger;
-import org.hawkular.alerts.api.services.ActionsService;
 import org.hawkular.alerts.api.services.AlertsCriteria;
 import org.hawkular.alerts.api.services.AlertsService;
-import org.hawkular.alerts.api.services.DefinitionsService;
 import org.hawkular.alerts.engine.log.MsgLogger;
-import org.hawkular.alerts.engine.rules.RulesEngine;
+import org.hawkular.alerts.engine.service.AlertsEngine;
 import org.jboss.logging.Logger;
 
 import com.google.gson.Gson;
@@ -72,92 +60,34 @@ import com.google.gson.GsonBuilder;
 public class DbAlertsServiceImpl implements AlertsService {
     private final MsgLogger msgLog = MsgLogger.LOGGER;
     private final Logger log = Logger.getLogger(DbAlertsServiceImpl.class);
-    private static final int DELAY;
-    private static final int PERIOD;
-
-    private final List<Data> pendingData;
-    private final List<Alert> alerts;
-    private final Set<Dampening> pendingTimeouts;
-    private final Map<Trigger, List<Set<ConditionEval>>> autoResolvedTriggers;
-    private final Set<Trigger> disabledTriggers;
-
-    private final Timer wakeUpTimer;
-    private TimerTask rulesTask;
 
     private Gson gson;
     private final String DS_NAME;
     private DataSource ds;
 
     @EJB
-    RulesEngine rules;
-
-    @EJB
-    DefinitionsService definitions;
-
-    @EJB
-    ActionsService actions;
-
-    /*
-        Init properties
-     */
-    static {
-        String sDelay = System.getProperty("org.hawkular.alerts.engine.DELAY");
-        String sPeriod = System.getProperty("org.hawkular.alerts.engine.PERIOD");
-        int dDelay = 1000;
-        int dPeriod = 2000;
-        try {
-            dDelay = Integer.parseInt(sDelay);
-            dPeriod = Integer.parseInt(sPeriod);
-        } catch (Exception ignored) {
-        }
-        DELAY = dDelay;
-        PERIOD = dPeriod;
-    }
+    AlertsEngine alertsEngine;
 
     public DbAlertsServiceImpl() {
         log.debugf("Creating instance.");
-        pendingData = new CopyOnWriteArrayList<>();
-        alerts = new CopyOnWriteArrayList<>();
-        pendingTimeouts = new CopyOnWriteArraySet<>();
-        autoResolvedTriggers = new HashMap<>();
-        disabledTriggers = new CopyOnWriteArraySet<>();
-        wakeUpTimer = new Timer("DbAlertsServiceImpl-Timer");
 
         DS_NAME = System.getProperty("org.hawkular.alerts.engine.datasource", "java:jboss/datasources/HawkularDS");
     }
 
-    @SuppressWarnings("unused")
-    public ActionsService getActions() {
-        return actions;
-    }
-
-    public void setActions(ActionsService actions) {
-        this.actions = actions;
-    }
-
-    public DefinitionsService getDefinitions() {
-        return definitions;
-    }
-
-    public void setDefinitions(DefinitionsService definitions) {
-        this.definitions = definitions;
-    }
-
-    public RulesEngine getRules() {
-        return rules;
-    }
-
-    public void setRules(RulesEngine rules) {
-        this.rules = rules;
-    }
-
-    @SuppressWarnings("unused")
-    public DataSource getDatasource() {
+    public DataSource getDs() {
         return ds;
     }
 
-    public void setDatasource(DataSource ds) {
+    public void setDs(DataSource ds) {
         this.ds = ds;
+    }
+
+    public AlertsEngine getAlertsEngine() {
+        return alertsEngine;
+    }
+
+    public void setAlertsEngine(AlertsEngine alertsEngine) {
+        this.alertsEngine = alertsEngine;
     }
 
     @PostConstruct
@@ -175,8 +105,6 @@ public class DbAlertsServiceImpl implements AlertsService {
                 msgLog.errorCannotConnectWithDatasource(e.getMessage());
             }
         }
-
-        reload();
     }
 
     @Override
@@ -442,244 +370,6 @@ public class DbAlertsServiceImpl implements AlertsService {
     }
 
     @Override
-    public void clear() {
-        rulesTask.cancel();
-
-        rules.clear();
-
-        pendingData.clear();
-        alerts.clear();
-        pendingTimeouts.clear();
-        autoResolvedTriggers.clear();
-        disabledTriggers.clear();
-
-        rulesTask = new RulesInvoker();
-        wakeUpTimer.schedule(rulesTask, DELAY, PERIOD);
-    }
-
-    @Override
-    public void reload() {
-        rules.reset();
-        if (rulesTask != null) {
-            rulesTask.cancel();
-        }
-
-        Collection<Trigger> triggers = null;
-        try {
-            triggers = definitions.getAllTriggers();
-        } catch (Exception e) {
-            log.debugf(e.getMessage(), e);
-            msgLog.errorDefinitionsService("Triggers", e.getMessage());
-        }
-        if (triggers != null && !triggers.isEmpty()) {
-            triggers.stream().filter(Trigger::isEnabled).forEach(this::reloadTrigger);
-        }
-
-        rules.addGlobal("log", log);
-        rules.addGlobal("actions", actions);
-        rules.addGlobal("alerts", alerts);
-        rules.addGlobal("pendingTimeouts", pendingTimeouts);
-        rules.addGlobal("autoResolvedTriggers", autoResolvedTriggers);
-        rules.addGlobal("disabledTriggers", disabledTriggers);
-
-        rulesTask = new RulesInvoker();
-        wakeUpTimer.schedule(rulesTask, DELAY, PERIOD);
-    }
-
-    @Override
-    public void reloadTrigger(final String tenantId, final String triggerId) {
-        if (isEmpty(tenantId)) {
-            throw new IllegalArgumentException("TenantId must be not null");
-        }
-        if (isEmpty(triggerId)) {
-            throw new IllegalArgumentException("TriggerId must be not null");
-        }
-
-        Trigger trigger = null;
-        try {
-            trigger = definitions.getTrigger(tenantId, triggerId);
-        } catch (Exception e) {
-            log.debugf(e.getMessage(), e);
-            msgLog.errorDefinitionsService("Trigger", e.getMessage());
-        }
-        if (null == trigger) {
-            log.debugf("Trigger not found for triggerId [" + triggerId + "], removing from rulebase if it exists");
-            Trigger doomedTrigger = new Trigger(triggerId, "doomed");
-            removeTrigger(doomedTrigger);
-            return;
-        }
-
-        reloadTrigger(trigger);
-    }
-
-    private void reloadTrigger(Trigger trigger) {
-        if (null == trigger) {
-            throw new IllegalArgumentException("Trigger must be not null");
-        }
-
-        // Look for the Trigger in the rules engine, if it is there then remove everything about it
-        removeTrigger(trigger);
-
-        if (trigger.isEnabled()) {
-            try {
-                Collection<Condition> conditionSet = definitions.getTriggerConditions(trigger.getTenantId(),
-                        trigger.getId(), null);
-                Collection<Dampening> dampenings = definitions.getTriggerDampenings(trigger.getTenantId(),
-                        trigger.getId(), null);
-
-                rules.addFact(trigger);
-                rules.addFacts(conditionSet);
-                if (!dampenings.isEmpty()) {
-                    rules.addFacts(dampenings);
-                }
-            } catch (Exception e) {
-                log.debugf(e.getMessage(), e);
-                msgLog.errorDefinitionsService("Conditions/Dampening", e.getMessage());
-            }
-        }
-    }
-
-    private void removeTrigger(Trigger trigger) {
-        if (null != rules.getFact(trigger)) {
-            // First remove the related Trigger facts from the engine
-            rules.removeFact(trigger);
-
-            // then remove everything else.
-            // We may want to do this with rules, because as is, we need to loop through every Fact in
-            // the rules engine doing a relatively slow check.
-            final String triggerId = trigger.getId();
-            rules.removeFacts(t -> {
-                if (t instanceof Dampening) {
-                    return ((Dampening) t).getTriggerId().equals(triggerId);
-                } else if (t instanceof Condition) {
-                    return ((Condition) t).getTriggerId().equals(triggerId);
-                }
-                return false;
-            });
-        } else {
-            log.debugf("Trigger not found. Not removed from rulebase %s", trigger);
-        }
-    }
-
-    @Override
-    public void sendData(Collection<Data> data) {
-        if (data == null) {
-            throw new IllegalArgumentException("Data must be not null");
-        }
-        pendingData.addAll(data);
-    }
-
-    @Override
-    public void sendData(Data data) {
-        if (data == null) {
-            throw new IllegalArgumentException("Data must be not null");
-        }
-        pendingData.add(data);
-    }
-
-    private class RulesInvoker extends TimerTask {
-        @Override
-        public void run() {
-            int numTimeouts = checkPendingTimeouts();
-            if (!pendingData.isEmpty() || numTimeouts > 0) {
-
-                log.debugf("Executing rules engine on [%1d] datums and [%2d] dampening timeouts.", pendingData.size(),
-                        numTimeouts);
-
-                try {
-                    if (pendingData.isEmpty()) {
-                        rules.fireNoData();
-
-                    } else {
-                        rules.addData(pendingData);
-                        pendingData.clear();
-                    }
-
-                    rules.fire();
-                    addAlerts(alerts);
-                    alerts.clear();
-                    handleDisabledTriggers();
-                    handleAutoResolvedTriggers();
-
-                } catch (Exception e) {
-                    e.printStackTrace();
-                    log.debugf("Error on rules processing: " + e);
-                    msgLog.errorProcessingRules(e.getMessage());
-                } finally {
-                    alerts.clear();
-                }
-            }
-        }
-
-        private int checkPendingTimeouts() {
-            if (pendingTimeouts.isEmpty()) {
-                return 0;
-            }
-
-            long now = System.currentTimeMillis();
-            Set<Dampening> timeouts = null;
-            for (Dampening d : pendingTimeouts) {
-                if (now < d.getTrueEvalsStartTime() + d.getEvalTimeSetting()) {
-                    continue;
-                }
-
-                d.setSatisfied(true);
-                try {
-                    log.debugf("Dampening Timeout Hit! %s", d.toString());
-                    rules.updateFact(d);
-                    if (null == timeouts) {
-                        timeouts = new HashSet<>();
-                    }
-                    timeouts.add(d);
-                } catch (Exception e) {
-                    log.error("Unable to update Dampening Fact on Timeout! " + d.toString(), e);
-                }
-
-            }
-
-            if (null == timeouts) {
-                return 0;
-            }
-
-            pendingTimeouts.removeAll(timeouts);
-            return timeouts.size();
-        }
-    }
-
-    private void handleDisabledTriggers() {
-        try {
-            for (Trigger t : disabledTriggers) {
-                try {
-                    t.setEnabled(false);
-                    definitions.updateTrigger(t.getTenantId(), t);
-
-                } catch (Exception e) {
-                    log.errorf("Failed to persist updated trigger. Could not autoDisable %s", t);
-                }
-            }
-        } finally {
-            disabledTriggers.clear();
-        }
-    }
-
-    private void handleAutoResolvedTriggers() {
-        try {
-            for (Map.Entry<Trigger, List<Set<ConditionEval>>> entry : autoResolvedTriggers.entrySet()) {
-                Trigger t = entry.getKey();
-                try {
-                    if (t.isAutoResolveAlerts()) {
-                        resolveAlertsForTrigger(t.getTenantId(), t.getId(), "AUTO", null, entry.getValue());
-                    }
-                } catch (Exception e) {
-                    log.errorf("Failed to resolve Alerts. Could not AutoResolve alerts for trigger %s", t);
-                }
-            }
-        } finally {
-            autoResolvedTriggers.clear();
-        }
-    }
-
-    @Override
     public void ackAlerts(String tenantId, Collection<String> alertIds, String ackBy, String ackNotes)
             throws Exception {
         if (isEmpty(tenantId)) {
@@ -786,6 +476,16 @@ public class DbAlertsServiceImpl implements AlertsService {
         }
 
         return alert;
+    }
+
+    @Override
+    public void sendData(Data data) throws Exception {
+        alertsEngine.sendData(data);
+    }
+
+    @Override
+    public void sendData(Collection<Data> data) throws Exception {
+        alertsEngine.sendData(data);
     }
 
 }
